Move HasBatchSize() and HasMaxIterations() from the callbacks to the functions traits.
